pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argilla-haystack"
description = "Argilla-Haystack Integration"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["annotation", "llm", "monitoring"]
authors = [{ name = "Argilla", email = "admin@argilla.io" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "argilla >= 1.18.0",
    "packaging >= 23.2",
    "typing-extensions >= 4.3.0"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["black == 23.10.0", "ruff == 0.1.0", "pre-commit >= 3.5.0"]
tests = ["pytest >= 7.4.0"]
haystack-v1 = ["farm-haystack['inference'] >= 1.23.0"]
haystack-v2 = ["haystack-ai >= 2.0.0a1"]

[project.urls]
Documentation = "https://github.com/argilla-io/argilla-haystack"
Issues = "https://github.com/argilla-io/argilla-haystack/issues"
Source = "https://github.com/argilla-io/argilla-haystack"

[tool.hatch.version]
path = "src/argilla_haystack/__init__.py"

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "C", "B"]
ignore = ["E501", "B905", "B008"]
exclude = ["docs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
